+++ src/main/java/com/dynatrace/openkit/protocol/Beacon.java
@@ -58,7 +58,6 @@ public class Beacon {
     private static final String BEACON_KEY_PROTOCOL_VERSION = "vv";
     private static final String BEACON_KEY_OPENKIT_VERSION = "va";
     private static final String BEACON_KEY_APPLICATION_ID = "ap";
-    private static final String BEACON_KEY_APPLICATION_NAME = "an";
     private static final String BEACON_KEY_APPLICATION_VERSION = "vn";
     private static final String BEACON_KEY_PLATFORM_TYPE = "pt";
     private static final String BEACON_KEY_AGENT_TECHNOLOGY_TYPE = "tt";
@@ -1064,7 +1063,6 @@ public class Beacon {
         addKeyValuePair(basicBeaconBuilder, BEACON_KEY_PROTOCOL_VERSION, ProtocolConstants.PROTOCOL_VERSION);
         addKeyValuePair(basicBeaconBuilder, BEACON_KEY_OPENKIT_VERSION, ProtocolConstants.OPENKIT_VERSION);
         addKeyValuePair(basicBeaconBuilder, BEACON_KEY_APPLICATION_ID, openKitConfiguration.getApplicationID());
-        addKeyValuePairIfNotNull(basicBeaconBuilder, BEACON_KEY_APPLICATION_NAME, openKitConfiguration.getApplicationName());
         addKeyValuePairIfNotNull(basicBeaconBuilder, BEACON_KEY_APPLICATION_VERSION, openKitConfiguration.getApplicationVersion());
         addKeyValuePair(basicBeaconBuilder, BEACON_KEY_PLATFORM_TYPE, ProtocolConstants.PLATFORM_TYPE_OPENKIT);
         addKeyValuePair(basicBeaconBuilder, BEACON_KEY_AGENT_TECHNOLOGY_TYPE, ProtocolConstants.AGENT_TECHNOLOGY_TYPE);