+++ src/test/java/com/dynatrace/openkit/core/objects/NullActionTest.java
@@ -116,19 +116,6 @@ public class NullActionTest {
     }
 
     @Test
-    public void deprecatedReportErrorReturnsSelf() {
-        // given
-        NullAction target = createNullAction();
-
-        // when
-        Action obtained = target.reportError("error name", 1337, "something bad");
-
-        // then
-        assertThat(obtained, is(instanceOf(NullAction.class)));
-        assertThat((NullAction)obtained, is(sameInstance(target)));
-    }
-
-    @Test
     public void reportErrorReturnsSelf() {
         // given
         NullAction target = createNullAction();