+++ src/test/java/com/dynatrace/openkit/core/objects/WebRequestTracerBaseImplTest.java
@@ -110,18 +110,6 @@ public class WebRequestTracerBaseImplTest {
     }
 
     @Test
-    public void aWebRequestTracerIsStoppedAfterStopHasBeenCalled() {
-        // given
-        WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
-
-        // when calling the stop method
-        target.stop();
-
-        // then
-        assertThat(target.isStopped(), is(true));
-    }
-
-    @Test
     public void aWebRequestTracerIsStoppedWithResponseCodeAfterStopHasBeenCalled() {
         // given
         WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
@@ -134,20 +122,6 @@ public class WebRequestTracerBaseImplTest {
     }
 
     @Test
-    public void setResponseCodeSetsTheResponseCode() {
-
-        // given
-        WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
-
-        // when setting response code
-        WebRequestTracer obtained = target.setResponseCode(418);
-
-        // then
-        assertThat(target.getResponseCode(), is(418));
-        assertThat(obtained, is(sameInstance((WebRequestTracer)target)));
-    }
-
-    @Test
     public void stopWithResponseCodeSetsTheResponseCode() {
         // given
         WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
@@ -157,21 +131,6 @@ public class WebRequestTracerBaseImplTest {
 
         // then
         assertThat(target.getResponseCode(), is(418));
-    }
-
-    @Test
-    public void setResponseCodeDoesNotSetTheResponseCodeIfStopped() {
-
-        // given
-        WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
-        target.stop();
-
-        // when setting response code
-        WebRequestTracer obtained = target.setResponseCode(418);
-
-        // then
-        assertThat(target.getResponseCode(), is(-1));
-        assertThat(obtained, is(sameInstance((WebRequestTracer)target)));
     }
 
     @Test
@@ -200,21 +159,6 @@ public class WebRequestTracerBaseImplTest {
 
         // then
         assertThat(target.getBytesSent(), is(1234));
-        assertThat(obtained, is(sameInstance((WebRequestTracer)target)));
-    }
-
-    @Test
-    public void setBytesSentDoesNotSetAnythingIfStopped() {
-
-        // given
-        WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
-        target.stop();
-
-        // when setting the sent bytes
-        WebRequestTracer obtained = target.setBytesSent(1234);
-
-        // then
-        assertThat(target.getBytesSent(), is(-1));
         assertThat(obtained, is(sameInstance((WebRequestTracer)target)));
     }
 
@@ -248,21 +192,6 @@ public class WebRequestTracerBaseImplTest {
     }
 
     @Test
-    public void setBytesReceivedDoesNotSetAnythingIfStopped() {
-
-        // given
-        WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
-        target.stop();
-
-        // when setting the received bytes
-        WebRequestTracer obtained = target.setBytesReceived(4321);
-
-        // then
-        assertThat(target.getBytesReceived(), is(-1));
-        assertThat(obtained, is(sameInstance((WebRequestTracer)target)));
-    }
-
-    @Test
     public void setBytesReceivedDoesNotSetAnythingIfStoppedWithResponseCode() {
         // given
         WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
@@ -292,22 +221,6 @@ public class WebRequestTracerBaseImplTest {
     }
 
     @Test
-    public void startDoesNothingIfAlreadyStopped() {
-
-        // given
-        WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
-        when(mockBeacon.getCurrentTimestamp()).thenReturn(123456789L);
-        target.stop();
-
-        // when starting web request tracing
-        WebRequestTracer obtained = target.start();
-
-        // then
-        assertThat(target.getStartTime(), is(0L));
-        assertThat(obtained, is(sameInstance((WebRequestTracer)target)));
-    }
-
-    @Test
     public void startDoesNothingIfAlreadyStoppedWithResponseCode() {
 
         // given
@@ -321,29 +234,6 @@ public class WebRequestTracerBaseImplTest {
         // then
         assertThat(target.getStartTime(), is(0L));
         assertThat(obtained, is(sameInstance((WebRequestTracer)target)));
-    }
-
-    @Test
-    public void stopCanOnlyBeExecutedOnce() {
-
-        // given
-        WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
-        when(mockBeacon.createSequenceNumber()).thenReturn(42);
-
-        // when executed the first time
-        target.stop();
-
-        // then
-        assertThat(target.getEndSequenceNo(), is(42));
-        verify(mockBeacon, times(2)).createSequenceNumber();
-        verify(mockBeacon, times(1)).addWebRequest(0, target);
-
-        // and when executed the second time
-        target.stop();
-
-        // then
-        verify(mockBeacon, times(2)).createSequenceNumber();
-        verify(mockBeacon, times(1)).addWebRequest(0, target);
     }
 
     @Test
@@ -382,35 +272,6 @@ public class WebRequestTracerBaseImplTest {
         verify(parentOpenKitObject, times(1)).onChildClosed(target);
         verify(parentOpenKitObject, times(1)).getActionID();
         verifyNoMoreInteractions(parentOpenKitObject);
-    }
-
-    @Test
-    public void stopNotifiesParentAndDetachesFromParent() {
-        WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
-        when(mockBeacon.createSequenceNumber()).thenReturn(42);
-
-        // when
-        target.stop();
-
-        // then
-        assertThat(target.getParent(), is(nullValue()));
-        verify(parentOpenKitObject, times(1)).onChildClosed(target);
-        verify(parentOpenKitObject, times(1)).getActionID();
-        verifyNoMoreInteractions(parentOpenKitObject);
-    }
-
-    @Test
-    public void stopUsesSetResponseCode() {
-        // given
-        WebRequestTracerBaseImpl target = new TestWebRequestTracerBaseImpl(logger, parentOpenKitObject, mockBeacon);
-
-        // when setting response code and stopping
-        WebRequestTracer obtained = target.setResponseCode(418);
-        target.stop();
-
-        // then
-        assertThat(target.getResponseCode(), is(418));
-        assertThat(obtained, is(sameInstance((WebRequestTracer)target)));
     }
 
     @Test